requirements.txt
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
get-bittrex-tick==1.0
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.4.4
pytz==2019.3
requests==2.23.0
six==1.14.0
typed-ast==1.4.1
tzlocal==2.0.0
urllib3==1.25.8
wrapt==1.11.2